♻️refactor : refactor project structure
main -> app -> router -> page to main -> app -> page for optimization.
